app/requirements.txt
websockets==12.0
python-dotenv==1.0.0
pypresence==4.2.1
